Hide fieldsets with only hidden fields
If all of the fields in a fieldset are hidden, ignore that fieldset.

https://github.com/azzlack/sanity-plugin-tabs/issues/31

---

I removed the check for if the length is greater than 0.

If it's an empty array, then after filtering and sorting it will still
be an empty array. Which is the initial value of `fieldSets`, so no
change.